mcp-llm-server/pyproject.toml
[project]
name = "mcp-llm-server"
version = "0.1.0"
description = "MCP Server for LLM operations - LangChain Gemini, injection guard, Korean NLP"
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = { text = "MIT" }
authors = [
    { name = "kapu" }
]
keywords = ["mcp", "llm", "langchain", "gemini", "korean-nlp", "injection-guard"]

dependencies = [
    "mcp>=1.22.0",
    "langchain-google-genai>=3.2.0",
    "langchain-core>=1.1.2",
    "langgraph>=0.2.0",
    "google-ai-generativelanguage>=0.6.0",
    # Latest available on PyPI is 0.2.1; pinned to avoid resolver failure
    "langgraph-checkpoint-redis==0.2.1",
    "kiwipiepy>=0.22.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "pyahocorasick>=2.2.0",
    "python-dotenv>=1.2.1",
    "cachetools>=6.2.2",
    "loguru>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "hypercorn>=0.17.0",
    "h2>=4.1.0",
    "asyncpg>=0.30.0",
    "orjson>=3.10.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "inline-snapshot>=0.18.0",
    "ruff>=0.8.0",
    "black>=24.0.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.0",
    "types-cachetools>=5.0.0",
]

[project.scripts]
# HTTP REST API 서버 (메인 진입점)
mcp-llm-server = "mcp_llm_server.http_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_llm_server"]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]
exclude = [".venv", "__pycache__", ".git"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
extend-select = [
    "RET",     # flake8-return
    "PLR2004", # magic numbers in comparisons
    "T20",     # no print
    "ANN",     # type annotations
    "BLE",     # bare except
    "PLR0913", # too many arguments
    "PLR0915", # too many statements
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # function call in default argument (DI-style patterns)
    "B905",   # zip strict
    "ARG001", # unused function argument (callback signatures)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "ANN001", "ANN201", "ANN202", "ANN204"]
"examples/**" = ["T201", "ANN201"]
"scripts/**" = ["T201", "ANN201", "BLE001"]

[tool.ruff.lint.isort]
known-first-party = ["mcp_llm_server"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-p",
    "no:langsmith",
    "-p",
    "no:langsmith.pytest_plugin",
    "-p",
    "no:langsmith_plugin",
    "-p",
    "no:pytest_langsmith",
]
filterwarnings = [
    "error",
    "ignore:Core Pydantic V1 functionality isn't compatible with Python 3.14 or greater.:UserWarning",
    "ignore:ForwardRef._evaluate is a private API.*:DeprecationWarning",
]

[tool.black]
target-version = ["py313"]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  \.venv
  | \.git
  | __pycache__
)/
'''

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_ignores = true

# Any 타입 정책 (실용적 접근)
# MCP/JSON 프로토콜 특성상 dict[str, Any] 불가피하지만 반환 Any는 경고 유지
disallow_any_generics = true      # 제네릭 Any 금지
disallow_any_unimported = true    # 미타입 외부 Any 금지
warn_return_any = true            # Any 반환 경고
implicit_reexport = false         # 암묵적 re-export 금지

# 타입 정의 필수
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

exclude = [
    "^.venv/",
    "^tests/",
]

# 외부 라이브러리 (스텁 불완전, 반환 Any 가능)
[[tool.mypy.overrides]]
module = [
    "kiwipiepy.*",
    "ahocorasick.*",
    "langchain_google_genai.*",
    "langchain_core.*",
    "langgraph.*",
    "langgraph_checkpoint_redis.*",
    "mcp.*",
    "pydantic.*",
    "fastapi.*",
    "uvicorn.*",
    "asyncpg.*",
]
ignore_missing_imports = true
disallow_any_expr = false
warn_return_any = true
